refactor: fix clippy warnings and rustfmt config
- storage_factory: use #[derive(Default)] with #[default] attribute
- l1_cache: collapse nested if statements into single condition
- dbnexus_ban_storage: remove unnecessary i64 type cast
- rustfmt: disable nightly-only format_strings option

+++ src/adapters/dbnexus_ban_storage.rs
@@ -149,7 +149,7 @@ impl BanStorage for DBNexusBanStorageAdapter {
             .expires_at
             .signed_duration_since(record.banned_at)
             .num_seconds()
-            .max(0) as i64;
+            .max(0);
 
         let model_id = existing.as_ref().map(|m| m.id).unwrap_or(0);
         let model = BanRecordModel {
